Partial work on moving WindowEvent->Event conversion to poll_event handler (gave up)
